fix task1.js and task2,js
using parseFloat() instead of parseInt()

+++ FL11_HW6/homework/js/task2.js
@@ -2,9 +2,9 @@ let a = prompt('Enter the length for side A of the triangle', '');
 let b = prompt('Enter the length for side B of the triangle', '');
 let c = prompt('Enter the length for side C of the triangle', '');
 
-if (parseInt(a) < parseInt(b) + parseInt(c) &&
-	parseInt(b) < parseInt(a) + parseInt(c) &&
-	parseInt(c) < parseInt(b) + parseInt(a)) {
+if (parseFloat(a) < parseFloat(b) + parseFloat(c) &&
+	parseFloat(b) < parseFloat(a) + parseFloat(c) &&
+	parseFloat(c) < parseFloat(b) + parseFloat(a)) {
 	if (a===b && b===c && c===a) {
 		console.log('Equivalent triangle');
 	} else {
